build: Add release-please workflow (#36)
In order to automate Changelog and releases we are going to integrate
release-please-action. It will require us to write the pull request and
commits. The best merge strategy where everything work as expected
should be squash-merge. With that strategy you can rewrite comments by
updating the pull request body -
https://github.com/googleapis/release-please#how-can-i-fix-release-notes.
The prefixes should be one of those: "feat", "fix", and "deps", "chore",
"build". "chore" and build" or any other prefix does not trigger
release. If the prefix ends with `!` it will count as breaking change
and will bump the major version.