chore(UX-1384): Proposal for theme overrides (#106)
This PR is a proposal of how we _could_ allow for overriding of the
light | dark color them and the regular | more constrast theme.
This implementation does not introduce any breaking changes.

To force override a color theme, we can add `data-theme: light` or
`data-theme: dark` to any element.
To force override a contrast theme, we can add `data-contrast: less` or
`data-contrast: more` to any element.

The drawback of this approach is that now we have **8** generated style
sheets, 4 primitives and 4 semantics. This could be reduced but for now
I have left it at 8 for readability.

@benken @DE7924 @aygurs Please give feedback on this approach

---------

Co-authored-by: github-actions <github-actions@github.com>